ai-pipeline/requirements.txt
fastapi>=0.115.6
uvicorn>=0.34.0
langgraph>=0.2.60
langchain>=0.1.4
langchain-core>=0.3.29
langchain-community>=0.3.14
langchain-groq>=0.1.6
python-dotenv>=1.0.1
openai>=1.46.0
groq>=0.11.0
google-generativeai>=0.8.2
pydantic>=2.7.0
tiktoken>=0.6.0
fastembed>=0.2.0 